[SwiftASTContext] Remove serialized options fallback.
There's a very complex logic in SwiftASTContext to deal with the
case the compiler options aren't serialized that look at
apple_flags in DWARF and tries to reconstruct the compile invocations.

This doesn't work, as it's been shown by every project that
disables serialization of debugging options and reports
that "debugging doesn't work".

We should just nuke this code.

<rdar://problem/55792430>